Adjust runtests so it can be invoked anywhere
No dependency on being in the root of the jboss-dockerfiles checkout
anymore.

Attempts to guess location of jboss-dockerfiles repo based on the
feature file supplied.

Presumes ce-ctf-tests is checked out *within* jboss-dockerfiles, as
the previous version of the script did, an as the instruction say to
do. (We will want to change this soon)

Work in a temporary directory and remove it after invocation.